feat: 처음 처리된 청크 시간정보 저장 및 조회 기능 추가
